fix(view): Another alternative view
This primarily is meant to be a simplification so its easier to add /
remove content.
We can evaluate fancier formatting, like tables, later when we know more
of what we want.

To simplify my work, I cut fields that I wasn't as sure of including.

I also wanted to take a small crack at some formatting changes so the
new, plain view was easier to read.

Since I was touching formatting, I went ahead and applied my indentation
suggestion.

This is by no means to say "this is THE design" but I think it helps
move us in that direction.